feat(ModelosBiblioteca): ajuste de modelos que se usaran en el proyecto

+++ src/main/java/com/sgb/biblioteca/model/CalculablePeso.java
@@ -1,5 +1,5 @@
 package com.sgb.biblioteca.model;
 
 public interface CalculablePeso {
-    Integer calcularPesoTotal();
+    Double calcularPesoTotal();
 }
